refactor: extract header related logic from ScanCodeDataProvider
